pyproject.toml
[project]
name = "decksmith"
version = "0.1.15"
description = "A command-line application to dynamically generate decks of cards from a JSON specification and a CSV data file, inspired by nandeck."
authors = [
    {name = "Julio Cabria", email = "juliocabria@tutanota.com"},
]
license = "GPL-2.0-only"
readme = "docs/README.md"
requires-python = ">=3.11"
dependencies = [
    "click",
    "crossfiledialog>=1.1.0",
    "flask>=3.1.2",
    "jval==1.0.6",
    "pandas",
    "pillow>=11.3.0",
    "pywin32>=311; sys_platform == 'win32'",
    "reportlab>=4.4.3",
    "ruamel-yaml>=0.19.1",
    "waitress>=3.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["decksmith*", "test*"]
namespaces = false

[project.scripts]
decksmith = "decksmith.main:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pylint>=4.0.4",
    "pynsist>=2.8",
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["build_nsis", "dist", "build", ".venv"]
